+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Include Scurrius pet name in pet notifications. (#410)
 - Bugfix: Update death safe exceptions on config import from clipboard. (#406)
 
 ## 1.8.2